requirements.txt
absl-py==2.1.0
accelerate==0.27.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bitsandbytes==0.39.1
black==23.7.0
blis==0.7.11
Brotli==1.0.9
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cloudpathlib==0.16.0
cmake==3.27.0
confection==0.1.4
cymem==2.0.8
datasets==2.14.1
decorator==5.1.1
deepspeed==0.13.0
dill==0.3.7
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
evaluate==0.4.1
executing==1.2.0
fairscale==0.4.13
filelock==3.12.2
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.42
hjson==3.1.0
huggingface-hub==0.20.2
hydra-core==1.3.2
idna==3.4
importlib-resources==6.1.1
inflate64==0.3.1
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lit==16.0.6
loralib==0.1.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multivolumefile==0.2.3
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
peft==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.9.1
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.25.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py7zr==0.20.5
pyarrow==12.0.1
pybcj==1.0.1
pycld2==0.41
pycryptodomex==3.18.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.15.1
pynvml==11.5.0
pyppmd==1.0.0
pysbd==0.3.4
python-dateutil==2.8.2
pytorch-triton==2.2.0+e28a256d71
pytz==2023.3
PyYAML==6.0.1
pyzstd==0.15.9
regex==2023.6.3
requests==2.31.0
responses==0.18.0
rouge-score==0.1.2
safetensors==0.3.1
scipy==1.10.1
sentencepiece==0.1.99
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==4.0.0
srsly==2.4.8
stack-data==0.6.2
sympy==1.12
termcolor==2.3.0
textblob==0.15.3
texttable==1.6.7
thinc==8.2.2
tokenize-rt==5.1.0
tokenizers==0.15.0
tomli==2.0.1
torch==2.3.0.dev20240119+cu121
tqdm==4.65.0
traitlets==5.9.0
transformers @ git+https://github.com/huggingface/transformers@3f69f415adcbdaedec154ba8eac220ef3276975d
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wandb==0.16.3
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.17.0
